backend/requirements.txt
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==3.1
itsdangerous==1.1.0
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
python-multipart==0.0.5
rfc3986==1.4.0
sniffio==1.2.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
python-dotenv
streamlit
SQLAlchemy
email-validator
psycopg2-binary